Move Date to an Int representation so it costs 4 instead of 8 bytes. It now has to be within -999_999 and 999_999 yeas instead of -99_999_999 and 99_999_999. I think nobody cares about such order of magnitude years. In theory it could hold more than 5 million years but this limit is more understandable.

+++ lib/src/commonMain/kotlin/maryk/lib/time/DateTime.kt
@@ -34,7 +34,7 @@ data class DateTime(
     fun toEpochMilli() = toEpochSecond() * MILLIS_PER_SECOND + time.milli
 
     /** Get the date time as the amount of seconds since 01-01-1970 */
-    fun toEpochSecond() = date.epochDay * SECONDS_PER_DAY + time.toSecondsOfDay()
+    fun toEpochSecond() = date.epochDay.toLong() * SECONDS_PER_DAY + time.toSecondsOfDay()
 
     companion object: IsTimeObject<DateTime>() {
         var MIN = DateTime(Date.MIN, Time.MIN)
@@ -48,7 +48,7 @@ data class DateTime(
 
         /** Create a DateTime by the amount of milliseconds since 01-01-1970 */
         fun ofEpochMilli(epochInMillis: Long): DateTime {
-            val epochDay = floor((epochInMillis / MILLIS_PER_DAY).toDouble()).toLong()
+            val epochDay = floor((epochInMillis / MILLIS_PER_DAY).toDouble()).toInt()
             val millisOfDay = floor((epochInMillis % MILLIS_PER_DAY).toDouble()).toInt()
             return DateTime(
                 Date.ofEpochDay(epochDay),
@@ -58,7 +58,7 @@ data class DateTime(
 
         /** Create a DateTime by the amount of seconds since 01-01-1970 */
         fun ofEpochSecond(epochInSeconds: Long, milli: Short = 0): DateTime {
-            val epochDay = floor(epochInSeconds.toDouble() / SECONDS_PER_DAY).toLong()
+            val epochDay = floor(epochInSeconds.toDouble() / SECONDS_PER_DAY).toInt()
             val secondOfDay = floor(epochInSeconds.toDouble() % SECONDS_PER_DAY).toInt()
             return DateTime(
                 Date.ofEpochDay(epochDay),
